+++ library/src/main/java/preference/PreferenceScreenActivity.java
@@ -1,0 +1,21 @@
+package preference;
+
+import android.os.Bundle;
+
+public class PreferenceScreenActivity extends WearPreferenceActivity {
+
+    static final String EXTRA_SCREEN = "preference_screen";
+
+    @Override protected void onCreate(final Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        if(!getIntent().hasExtra(EXTRA_SCREEN)) {
+            throw new IllegalStateException("Missing required extra EXTRA_SCREEN");
+        }
+
+        final WearPreferenceScreen screen = (WearPreferenceScreen) getIntent().getSerializableExtra(EXTRA_SCREEN);
+        setTitle(screen.getTitle(this));
+        addPreferencesFromPreferenceScreen(screen);
+    }
+
+}